move Configuration class to its own file
