+++ demoapp/src/main/java/me/uport/sdk/demoapp/DemoApplication.kt
@@ -1,6 +1,7 @@
 package me.uport.sdk.demoapp
 
 import android.app.Application
+import me.uport.sdk.Configuration
 import me.uport.sdk.Uport
 
 /**
@@ -11,7 +12,7 @@ class DemoApplication : Application() {
     override fun onCreate() {
         super.onCreate()
 
-        val config = Uport.Configuration()
+        val config = Configuration()
                 .setApplicationContext(this)
 
         Uport.initialize(config)
