+++ sdk/src/main/java/me/uport/sdk/Configuration.kt
@@ -1,0 +1,38 @@
+package me.uport.sdk
+
+import android.content.Context
+import me.uport.sdk.core.EthNetwork
+import me.uport.sdk.core.IFuelTokenProvider
+
+/**
+ * Encapsulates a configuration setup for the uPort SDK
+ */
+class Configuration {
+
+    lateinit var applicationContext: Context
+    lateinit var fuelTokenProvider: IFuelTokenProvider
+    lateinit var network: EthNetwork
+
+    @Suppress("unused")
+    fun setFuelTokenProvider(provider: IFuelTokenProvider): Configuration {
+        this.fuelTokenProvider = provider
+        return this
+    }
+
+    /**
+     * Plugs the uPort SDK with an application context.
+     */
+    fun setApplicationContext(context: Context): Configuration {
+        this.applicationContext = context.applicationContext
+        return this
+    }
+
+
+    /**
+     * Sets the default network to be used for Ethereum interactions
+     */
+    fun setEthNetwork(network: EthNetwork) {
+        this.network = network
+    }
+
+}
